+++ src/theme/index.ts
@@ -12,6 +12,7 @@ import Accordion from "./components/accordion";
 import Breadcrumb from "./components/breadcrumb";
 import Button from "./components/button";
 import Checkbox from "./components/checkbox";
+import ComponentWrapper from "./components/componentWrapper";
 import CustomCheckboxGroup from "./components/customCheckboxGroup";
 import CustomRadioGroup from "./components/customRadioGroup";
 import CustomSelect from "./components/select";
@@ -20,9 +21,11 @@ import Fieldset from "./components/fieldset";
 import Heading from "./components/heading";
 import HelperErrorText from "./components/helperErrorText";
 import Hero from "./components/hero";
+import HorizontalRule from "./components/horizontalRule";
 import Icon from "./components/icon";
 import Label from "./components/label";
 import Link from "./components/link";
+import List from "./components/list";
 import NotificationStyles from "./components/notification";
 import Radio from "./components/radio";
 import SearchBar from "./components/searchBar";
@@ -31,6 +34,7 @@ import Tabs from "./components/tabs";
 import TemplateStyles from "./components/template";
 import Text from "./components/text";
 import TextInput from "./components/textInput";
+import VideoPlayer from "./components/videoPlayer";
 
 /**
  * See Chakra default theme for shape of theme object:
@@ -67,6 +71,7 @@ const theme = extendTheme({
     Breadcrumb,
     Button,
     Checkbox,
+    ComponentWrapper,
     CustomCheckboxGroup,
     CustomRadioGroup,
     CustomSelect,
@@ -75,9 +80,11 @@ const theme = extendTheme({
     Heading,
     HelperErrorText,
     Hero,
+    HorizontalRule,
     Icon,
     Label,
     Link,
+    List,
     ...NotificationStyles,
     Radio,
     SearchBar,
@@ -86,13 +93,13 @@ const theme = extendTheme({
     ...TemplateStyles,
     Text,
     TextInput,
+    VideoPlayer,
   },
-  // Chakra prefixes its own CSS variables with `--chakra` by default but this
-  // can be updated to be anything we want. This can be "nypl" to have the
-  // `--nypl` prefix or even "" to have no prefix.
-  // config: {
-  //   cssVarPrefix: "nypl",
-  // },
+  // Use `cssVarPrefix` to set the prefix used on the CSS vars produced by
+  // Chakra. By default, Chakra prefixes its own CSS variables with `--chakra`.
+  config: {
+    cssVarPrefix: "nypl",
+  },
 });
 
 export default theme;